requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pyodbc
jinja2
python-multipart
